doc: improve typedocs; rename features->capabilities
