+++ .gitignore
@@ -1,6 +1,9 @@
 # test artifacts
 tests/artifacts
 
+# generated docs
+docs/
+
 # Logs
 logs
 *.log
