+++ justfile
@@ -82,9 +82,9 @@ _build out args='[]': prepare
     # bsd sed vs gnu sed: the former "-i" REQUIRES an argument, the latter
     # REQUIRES not having an argument
     find "dist/{{ out }}" -name '*.js' | if [ $(uname) == 'Darwin' ]; then
-      xargs -I{} sed -i '' -e '/^(ex|im)port/s/.ts"/.js"/g' '{}'
+      xargs -I{} sed -i '' -e '/^((ex|im)port|} from)/s/.ts"/.js"/g' '{}'
     else
-      xargs -I{} sed -i -e '/^(ex|im)port/s/.ts"/.js"/g' '{}'
+      xargs -I{} sed -i -e '/^((ex|im)port|} from)/s/.ts"/.js"/g' '{}'
     fi
 
     # build types (TODO: switch module target based on incoming args)
@@ -119,7 +119,7 @@ build_worker_node out='worker/node' args='[]':
       [{
         "platform": "node",
         "alias": {
-          "js-sdk:features": "./src/polyfills/node-features.ts",
+          "js-sdk:capabilities": "./src/polyfills/node-capabilities.ts",
           "js-sdk:fs": "node:fs",
           "js-sdk:wasi": "./src/polyfills/node-wasi.ts",
         }
@@ -133,7 +133,7 @@ build_worker_browser out='worker/browser' args='[]':
       [{
         "format": "esm",
         "alias": {
-          "js-sdk:features": "./src/polyfills/browser-features.ts",
+          "js-sdk:capabilities": "./src/polyfills/browser-capabilities.ts",
           "node:worker_threads": "./src/polyfills/worker-node-worker_threads.ts",
           "js-sdk:fs": "./src/polyfills/browser-fs.ts",
           "js-sdk:wasi": "./src/polyfills/browser-wasi.ts",
@@ -150,7 +150,7 @@ build_node_cjs out='cjs' args='[]':
         "platform": "node",
         "minify": false,
         "alias": {
-          "js-sdk:features": "./src/polyfills/node-features.ts",
+          "js-sdk:capabilities": "./src/polyfills/node-capabilities.ts",
           "js-sdk:worker-url": "./dist/worker/node/worker-url.ts",
           "js-sdk:fs": "node:fs/promises",
           "js-sdk:wasi": "./src/polyfills/node-wasi.ts",
@@ -180,7 +180,7 @@ build_node_esm out='esm' args='[]':
         "format": "esm",
         "minify": false,
         "alias": {
-          "js-sdk:features": "./src/polyfills/node-features.ts",
+          "js-sdk:capabilities": "./src/polyfills/node-capabilities.ts",
           "js-sdk:worker-url": "./dist/worker/node/worker-url.ts",
           "js-sdk:fs": "node:fs/promises",
           "js-sdk:wasi": "./src/polyfills/node-wasi.ts",
@@ -200,7 +200,7 @@ build_bun out='bun' args='[]':
         "minify": false,
         "alias": {
           "js-sdk:worker-url": "./src/polyfills/bun-worker-url.ts",
-          "js-sdk:features": "./src/polyfills/bun-features.ts",
+          "js-sdk:capabilities": "./src/polyfills/bun-capabilities.ts",
           "js-sdk:fs": "node:fs/promises",
           "js-sdk:wasi": "./src/polyfills/node-wasi.ts",
         }
@@ -218,7 +218,7 @@ build_browser out='browser' args='[]':
         "define": {"global": "window"},
         "format": "esm",
         "alias": {
-          "js-sdk:features": "./src/polyfills/browser-features.ts",
+          "js-sdk:capabilities": "./src/polyfills/browser-capabilities.ts",
           "node:worker_threads": "./src/polyfills/host-node-worker_threads.ts",
           "js-sdk:fs": "./src/polyfills/browser-fs.ts",
           "js-sdk:worker-url": "./dist/worker/browser/worker-url.ts",
@@ -332,7 +332,7 @@ serve-docs: docs
   python3 -m http.server 8000 -d docs/
 
 watch-docs: prepare
-  watchexec -r -w src -w README.md just serve-docs
+  watchexec -r -w types -w src -w README.md just serve-docs
 
 serve port='8124' logs='true':
     #!/usr/bin/env node
