+++ src/polyfills/browser-capabilities.ts
@@ -1,4 +1,6 @@
-export const FEATURES = {
+import type { Capabilities } from '../interfaces.ts';
+
+export const CAPABILITIES: Capabilities = {
   // When false, shared buffers have to be copied to an array
   // buffer before passing to Text{En,De}coding()
   allowSharedBufferCodec: false,
